Move compositor and base doc from lib to layout
