+++ app/assets/javascripts/app/layout/base_document.rb
