+++ app/assets/javascripts/app/layout/compositor.rb
